# docs/swagger.yaml
basePath: /
definitions:
  requests.TodoRequest:
    properties:
      assignee:
        type: string
      description:
        type: string
      priority:
        type: integer
      title:
        type: string
    required:
    - assignee
    - description
    - priority
    - title
    type: object
  response.response:
    properties:
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Todo REST API built with Go and Gin.
  title: Todo API
  version: "1.0"
paths:
  /api/v1/todo/create:
    post:
      consumes:
      - application/json
      description: Create a new todo item
      parameters:
      - description: Todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/requests.TodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new todo
      tags:
      - Todos
  /api/v1/todo/delete/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete a todo by UUID
      parameters:
      - description: Todo UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: Delete todo
      tags:
      - Todos
  /api/v1/todo/list:
    get:
      description: Get a paginated list of todos with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: "0"
        description: Todo status
        in: query
        name: status
        type: string
      - default: "1"
        description: Todo priority
        in: query
        name: priority
        type: string
      - description: Todo assignee
        in: query
        name: assignee
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: List todos
      tags:
      - Todos
swagger: "2.0"